--- ast-utilities-oo.h
@@ -1,21 +1,0 @@
-// Node Stubs
-class Identifier_Node;
-class Type_Node;
-class Formal_Node;
-class Def_Node;
-class Statement_Node;
-class Print_Node;
-class Body_Node;
-class Return_Node;
-class If_Node;
-class Not_Node;
-class Function_Call_Node;
-class Negate_Node;
-class Block_Node; 
-class Binary_Expression_Node;
-class Expr_Node_Type;
-class Number_Node;
-class Boolean_Node;
-class Operator_Node;
-class Program_Node;
-class xNode;
