+++ ast-utilities.cpp
@@ -1,39 +1,558 @@
 #include <string>
+#include <iostream>
+#include <sstream>
+#include "ast-utilities.hpp"
 
-// Simple nodes
-struct identifier_ast {std::string   value;};
-struct type_ast       {tokenType     value;};
-struct number_ast     {unsigned long value;};
-struct boolean_ast    {bool          value;};
-struct operator_ast   {char          value;};
+/**************************************
+           BASIC NODE TYPES
+**************************************/
 
-struct formal_ast  {
-  identifier_ast identifier;
-  type_ast       type;
+class IdentifierNode : public Node {
+  protected:
+    std::string name = "Identifier";
+    std::string value;
+  public:
+    IdentifierNode() {
+      value = "";
+    }
+    IdentifierNode(std::string val) {
+      value = val;
+    }
+    void print(int d) {
+      pad(d);
+      cout << name <<  ": " << value << '\n';
+    }
 };
 
-struct if-ast {
-  leftChild      then_ast;// then expression node
-  rightChild     else_ast;// else expression node
+class TypeNode : public Node {
+  protected:
+    std::string name = "Type";
+    std::string value;
+  public:
+    TypeNode() {
+      value = "";
+    }
+    TypeNode(tokenType val) {
+      if (val == booleanvalue) {
+        value = "boolean";
+      }
+      else if (val == number) {
+        value = "number";
+      }
+    }
+    void print(int d) {
+      pad(d);
+      cout << name << ": " << value << '\n';
+    }
 };
 
-struct then-ast {
-    leftChild    expr_ast;//expr with boolean result
-    rightChild   expr_ast;//expr to perform if L-child is true
+class BinaryOpNode : public Node {
+  protected:
+    std::string name = "Binary Operator";
+    char        value;
+  public:
+    BinaryOpNode(char val) {
+      value = val;
+    }
+    BinaryOpNode() {}
+    void print(int d) {
+      pad(d);
+      cout << name << ": " << value << '\n';
+    }
 };
 
-struct else-ast {
-    leftChild    expr_ast;//expr with boolean result
-    rightChild   expr_ast;//expr to perform if L-child is false
+class NumberNode : public Node {
+  protected:
+    std::string  name = "Number";
+    unsigned int value;
+  public:
+    NumberNode(std::string val) {
+      istringstream(val) >> value;
+    }
+    void print(int d) {
+      pad(d);
+      cout << name << ": " << value << '\n';
+    }
 };
 
-// Struct Prototype
-struct /*Name*/ {
-  //(Values)*
+class BooleanNode : public Node {
+  protected:
+    std::string name = "Boolean";
+    bool        value;
+  public:
+    BooleanNode(std::string val) {
+      if (val.front() == 't') {
+        value = true;
+      }
+      else {
+        value = false;
+      }
+    }
+    void print(int d) {
+      pad(d);
+      cout << name << ": ";
+      if (value) cout << "true"; else cout << "false";
+      cout << '\n';
+    }
 };
 
-// Binary Prototype
-struct /*name*/_ast {
-    leftChild    /*node*/_ast;
-    rightChild   /*node*/_ast;
+/***************************************
+            Complex nodes!
+***************************************/
+
+class ProgramNode : public Node {
+  protected:
+    std::string     name = "Program";
+    IdentifierNode  value;
+    FormalList     *formals;
+    DefList        *defs;
+    BodyNode       *body;
+  public:
+    void setValue(IdentifierNode *i) {value = *i;};
+    void setFormals(FormalList *i) {formals = i;};
+    void setDefs(DefList *i) {defs = i;};
+    void setBody(BodyNode *i) {body = i;};
+    ProgramNode(IdentifierNode *v, FormalList *f, DefList *d, BodyNode *b){
+      setValue(v);
+      setFormals(f);
+      setDefs(d);
+      setBody(b);
+    }
+    void print(int);
 };
+
+class FormalParamNode : public Node {
+  protected:
+    std::string    name = "Formal Parameter";
+    IdentifierNode value;
+    TypeNode       type;
+  public:
+    void setValue(IdentifierNode i) {value = i;}
+    void setType(TypeNode t) {type = t;}
+    FormalParamNode(IdentifierNode i, TypeNode t) {
+      setValue(i);
+      setType(t);
+    }
+    void print(int d) {
+      pad(d);
+      cout << name << '\n';
+      value.print(d+1);
+      type.print(d+1);
+    }
+};
+
+class FormalList : public Node {
+  protected:
+    std::string      name = "List of formal parameters";
+    FormalParamNode *value;
+    FormalList      *next;
+  public:
+    FormalList(FormalParamNode *f) {
+      value = f;
+    }
+    void addFormal(FormalParamNode f) {
+      if (next == NULL) {
+        next = new FormalList(&f);
+      }
+      else {
+        (*next).addFormal(f);
+      }
+    }
+    void print(int d) {
+      if (value != NULL) {(*value).print(d);}
+      if (next != NULL) {(*next).print(d);}
+    }
+};
+
+class DefNode : public Node {
+  protected:
+    std::string     name = "Definition";
+    IdentifierNode  value;
+    FormalList     *formals;
+    TypeNode        type;
+    BodyNode       *body;
+  public:
+    void setValue(IdentifierNode *v) {value = *v;};
+    void setFormals(FormalList *f) {formals = f;};
+    void setType(TypeNode *t) {type = *t;};
+    void setBody(BodyNode *b) {body = b;};
+    DefNode(IdentifierNode *v, FormalList *f, TypeNode *t, BodyNode *b) {
+      setValue(v);
+      setFormals(f);
+      setType(t);
+      setBody(b);
+    }
+    void print(int);
+};
+
+class DefList : public Node {
+  protected:
+    std::string  name = "Definitions List";
+    DefNode     *value;
+    DefList     *next;
+  public:
+    DefList(DefNode *d) {
+      value = d;
+    }
+    void addDef(DefNode d) {
+      if (next == NULL) {
+        next = new DefList(&d);
+      }
+      else {
+        (*next).addDef(d);
+      }
+    }
+    void print(int d) {
+      if (value != NULL) {(*value).print(d);}
+      if (next  != NULL) { (*next).print(d);}
+    }
+};
+
+class BodyNode : public Node {
+  protected:
+    std::string    name = "Body";
+    StatementList *value;
+  public:
+    BodyNode(StatementList *s) {
+      value = s;
+    }
+    void print(int);
+};
+
+enum StatementType {printstmt, returnstmt};
+
+class PrintStmtNode : public Node {
+  protected:
+    std::string     name = "Print Statement";
+    ExpressionNode *value;
+  public:
+    void setValue(ExpressionNode *e) {value = e;}
+    PrintStmtNode (ExpressionNode *e) {setValue(e);}
+    void print(int);
+};
+
+class ReturnStmtNode : public Node {
+  protected:
+    std::string     name = "Return Statement";
+    ExpressionNode *value;
+  public:
+    void setValue(ExpressionNode *e) {value = e;}
+    ReturnStmtNode (ExpressionNode *e) {setValue(e);}
+    void print(int);
+};
+
+class StatementNode : public Node {
+  protected:
+    std::string    name = "Statement";
+    StatementType  type;
+    void          *value;
+  public:
+    void setValue(PrintStmtNode *p) {
+      type  = printstmt;
+      value = p;
+    }
+    void setValue(ReturnStmtNode *r) {
+      type = returnstmt;
+      value = r;
+    }
+    StatementNode(PrintStmtNode  *p) {setValue(p);}
+    StatementNode(ReturnStmtNode *r) {setValue(r);}
+    void print(int d) {
+      switch(type) {
+        case printstmt:
+          (*static_cast<PrintStmtNode*>(value)).print(d);break;
+        case returnstmt:
+          (*static_cast<ReturnStmtNode*>(value)).print(d);break;
+      }
+    }
+};
+
+class StatementList : public Node {
+  protected:
+    std::string    name = "Statement List";
+    StatementNode *value;
+    StatementList *next;
+  public:
+    StatementList(StatementNode *s) {value = s;}
+    void addStatement(StatementNode s) {
+      if (next == NULL) {
+        next = new StatementList(&s);
+      }
+      else {
+        (*next).addStatement(s);
+      }
+    }
+    void print(int d) {
+      if (value != NULL) {(*value).print(d);}
+      if (next  != NULL) { (*next).print(d);}
+    }
+};
+
+class IfExprNode : public Node {
+  protected:
+    std::string     name = "If Expression";
+    ExpressionNode *test;
+    ExpressionNode *thenClause;
+    ExpressionNode *elseClause;
+  public:
+    void setTest(ExpressionNode *e) {test = e;}
+    void setThen(ExpressionNode *e) {thenClause = e;}
+    void setElse(ExpressionNode *e) {elseClause = e;}
+    IfExprNode (ExpressionNode *t, ExpressionNode *th, ExpressionNode *e) {
+      setTest(t);
+      setThen(th);
+      setElse(e);
+    }
+    void print(int);
+};
+
+class NotExprNode : public Node {
+  protected:
+    std::string     name = "Not Expression";
+    ExpressionNode *value;
+  public:
+    void setValue(ExpressionNode *e) {value = e;}
+    NotExprNode (ExpressionNode *e) {setValue(e);}
+    void print(int);
+};
+
+class FunctionCallNode : public Node {
+  protected:
+    std::string     name = "Function Call";
+    IdentifierNode  functionName;
+    ExpressionList *functionBody;
+  public:
+    void setName(IdentifierNode *i) {functionName = *i;}
+    void setBody(ExpressionList *e) {functionBody = e;}
+    FunctionCallNode(IdentifierNode *i, ExpressionList *e) {
+      setName(i);
+      setBody(e);
+    }
+    void print(int);
+};
+
+class NegateExprNode : public Node {
+  protected:
+    std::string     name = "Negate Expression";
+    ExpressionNode *value;
+  public:
+    void setValue(ExpressionNode *e) {value = e;}
+    NegateExprNode(ExpressionNode *e) {setValue(e);}
+    void print(int);
+};
+
+class BlockExprNode : public Node {
+  protected:
+    std::string     name = "Block Expression";
+    ExpressionNode *value;
+  public:
+    void setValue(ExpressionNode *e) {value = e;}
+    BlockExprNode(ExpressionNode *e) {setValue(e);}
+    void print(int);
+};
+
+class BinaryExprNode : public Node {
+  protected:
+    std::string     name = "Binary Expression";
+    ExpressionNode *left;
+    BinaryOpNode    op;
+    ExpressionNode *right;
+  public:
+    void setLeft (ExpressionNode *e) {left  = e;}
+    void setOp   (BinaryOpNode *o)   {op    = *o;}
+    void setRight(ExpressionNode *e) {right = e;}
+    BinaryExprNode(ExpressionNode *l, BinaryOpNode *o, ExpressionNode *r) {
+      setLeft(l);
+      setOp(o);
+      setRight(r);
+    }
+    void print(int);
+};
+
+/**************************************
+           Expression node!
+         (this one's a doozy)
+**************************************/
+
+enum ExpressionType {binaryexpr, ifexpr, notexpr, functioncall,
+identifierexpr, numberexpr, booleanexpr, negateexpr, blockexpr};
+
+class ExpressionNode : public Node {
+  protected:
+    std::string     name = "Expression";
+    ExpressionType  type;
+    void           *value;
+  public:
+    void setValue(BinaryExprNode *b) {
+      type = binaryexpr;
+      value = b;
+    }
+    void setValue(IfExprNode *i) {
+      type = ifexpr;
+      value = i;
+    }
+    void setValue(NotExprNode *n) {
+      type = notexpr;
+      value = n;
+    }
+    void setValue(FunctionCallNode *f) {
+      type = functioncall;
+      value = f;
+    }
+    void setValue(IdentifierNode *i) {
+      type = identifierexpr;
+      value = i;
+    }
+    void setValue(NumberNode *n) {
+      type = numberexpr;
+      value = n;
+    }
+    void setValue(BooleanNode *b) {
+      type = booleanexpr;
+      value = b;
+    }
+    void setValue(NegateExprNode *n) {
+      type = negateexpr;
+      value = n;
+    }
+    void setValue(BlockExprNode *b) {
+      type = blockexpr;
+      value = b;
+    }
+    ExpressionNode(BinaryExprNode   *e) {setValue(e);}
+    ExpressionNode(IfExprNode       *e) {setValue(e);}
+    ExpressionNode(NotExprNode      *e) {setValue(e);}
+    ExpressionNode(FunctionCallNode *e) {setValue(e);}
+    ExpressionNode(IdentifierNode   *e) {setValue(e);}
+    ExpressionNode(NumberNode       *e) {setValue(e);}
+    ExpressionNode(BooleanNode      *e) {setValue(e);}
+    ExpressionNode(NegateExprNode   *e) {setValue(e);}
+    ExpressionNode(BlockExprNode    *e) {setValue(e);}
+    void print(int d) {
+      switch(type) {
+        case binaryexpr:
+          (*static_cast<BinaryExprNode*>(value)).print(d);break;
+        case ifexpr:
+          (*static_cast<IfExprNode*>(value)).print(d);break;
+        case notexpr:
+          (*static_cast<NotExprNode*>(value)).print(d);break;
+        case functioncall:
+          (*static_cast<FunctionCallNode*>(value)).print(d);break;
+        case identifierexpr:
+          (*static_cast<IdentifierNode*>(value)).print(d);break;
+        case numberexpr:
+          (*static_cast<NumberNode*>(value)).print(d);break;
+        case booleanexpr:
+          (*static_cast<BooleanNode*>(value)).print(d);break;
+        case negateexpr:
+          (*static_cast<NegateExprNode*>(value)).print(d);break;
+        case blockexpr:
+          (*static_cast<BlockExprNode*>(value)).print(d);break;
+        default: break;
+      }
+    }
+};
+
+class ExpressionList : public Node {
+  protected:
+    std::string     name = "Expression List";
+    ExpressionNode *value;
+    ExpressionList *next;
+  public:
+    ExpressionList(ExpressionNode *e) {
+      value = e;
+    }
+    void addExpression(ExpressionNode e) {
+      if (next == NULL) {
+        next = new ExpressionList(&e);
+      }
+      else {
+        (*next).addExpression(e);
+      }
+    }
+    void print(int d) {
+      if (value != NULL) {(*value).print(d);}
+      if (next  != NULL) { (*next).print(d);}
+    }
+};
+
+/**************************************
+Print functions that have to be defined
+down here to avoid forward declarations
+**************************************/
+
+void BinaryExprNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*left ).print(d+1);
+  op.print(d+1);
+  (*right).print(d+1);
+}
+
+void BlockExprNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*value).print(d+1);
+}
+
+void NegateExprNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*value).print(d+1);
+}
+
+void FunctionCallNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  functionName.print(d+1);
+  (*functionBody).print(d+1);
+}
+
+void NotExprNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*value).print(d+1);
+}
+
+void IfExprNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*test).print(d+1);
+  (*thenClause).print(d+1);
+  (*elseClause).print(d+1);
+}
+
+void ReturnStmtNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*value).print(d+1);
+}
+
+void PrintStmtNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*value).print(d+1);
+}
+
+void BodyNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  (*value).print(d+1);
+}
+
+void DefNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  value.print(d+1);
+  if (formals != NULL) {(*formals).print(d+1);}
+  type.print(d+1);
+  (*body).print(d+1);
+}
+
+void ProgramNode::print(int d) {
+  pad(d);
+  cout << name << '\n';
+  if (formals != NULL) {(*formals).print(d+1);}
+  if (defs != NULL) {(*defs).print(d+1);}
+  (*body).print(d+1);
+}
